New function and fix js mode.
The new function is for any company that has an internal jira server.
It looks for something that looks like a jira ticket, and
then opens the url for the issue.  This is for stuff like
jira tickets in commit messages.

Also make sure we use 2 spaces for all of our JavaScript editing,
including jsx.