Change how do-cmd works (also add new functions).
Update the doc string for do-cmd and make it so it doesn't
used bash -c as an intermediary.  There really isn't a need for that.

This may come up as an issue again if I want to run the command in a
special environment.

Also add a few new assoc-* commands for getting keys or values.